+++ acmed/src/acme_proto/account.rs
@@ -4,11 +4,12 @@ use crate::certificate::Certificate;
 use crate::endpoint::Endpoint;
 use crate::jws::encode_jwk;
 use crate::storage;
-use acme_common::crypto::{gen_keypair, KeyPair};
+use acme_common::crypto::{gen_keypair, JwsSignatureAlgorithm, KeyPair};
 use acme_common::error::Error;
 
 pub struct AccountManager {
     pub key_pair: KeyPair,
+    pub signature_algorithm: JwsSignatureAlgorithm,
     pub account_url: String,
     pub orders_url: String,
 }
@@ -21,14 +22,18 @@ impl AccountManager {
     ) -> Result<Self, Error> {
         // TODO: store the key id (account url)
         let key_pair = storage::get_account_keypair(cert)?;
+        let signature_algorithm = endpoint.signature_algorithm;
         let kp_ref = &key_pair;
         let account = Account::new(cert, endpoint);
         let account = serde_json::to_string(&account)?;
         let acc_ref = &account;
-        let data_builder = |n: &str, url: &str| encode_jwk(kp_ref, acc_ref.as_bytes(), url, n);
+        let data_builder = |n: &str, url: &str| {
+            encode_jwk(kp_ref, &signature_algorithm, acc_ref.as_bytes(), url, n)
+        };
         let (acc_rep, account_url) = http::new_account(endpoint, root_certs, &data_builder)?;
         let ac = AccountManager {
             key_pair,
+            signature_algorithm,
             account_url,
             orders_url: acc_rep.orders.unwrap_or_default(),
         };
@@ -37,16 +42,21 @@ impl AccountManager {
     }
 }
 
-pub fn init_account(cert: &Certificate) -> Result<(), Error> {
+pub fn init_account(cert: &Certificate, endpoint: &Endpoint) -> Result<(), Error> {
     if !storage::account_files_exists(cert) {
-        // TODO: allow to change the account key type
-        let key_pair = gen_keypair(crate::DEFAULT_ACCOUNT_KEY_TYPE)?;
+        let key_pair = gen_keypair(endpoint.key_type)?;
         storage::set_account_keypair(cert, &key_pair)?;
         cert.info(&format!("Account {} created", &cert.account.name));
     } else {
-        // TODO: check if the keys are suitable for the specified signature algorithm
-        // and, if not, initiate a key rollover.
-        cert.debug(&format!("Account {} already exists", &cert.account.name));
+        let key_pair = storage::get_account_keypair(cert)?;
+        if key_pair.key_type != endpoint.key_type {
+            cert.debug(&format!("Account {} has a key pair of type {} while {} was expected. Creating a new {} key pair.", &cert.account.name, key_pair.key_type, endpoint.key_type, endpoint.key_type));
+            // TODO: Do a propper key rollover?
+            let key_pair = gen_keypair(endpoint.key_type)?;
+            storage::set_account_keypair(cert, &key_pair)?;
+        } else {
+            cert.debug(&format!("Account {} already exists", &cert.account.name));
+        }
     }
     Ok(())
 }