+++ acmed/src/endpoint.rs
@@ -1,4 +1,5 @@
 use crate::acme_proto::structs::Directory;
+use acme_common::crypto::{JwsSignatureAlgorithm, KeyType};
 use acme_common::error::Error;
 use nom::bytes::complete::take_while_m_n;
 use nom::character::complete::digit1;
@@ -6,6 +7,7 @@ use nom::combinator::map_res;
 use nom::multi::fold_many1;
 use nom::IResult;
 use std::cmp;
+use std::str::FromStr;
 use std::thread;
 use std::time::{Duration, Instant};
 
@@ -17,6 +19,8 @@ pub struct Endpoint {
     pub nonce: Option<String>,
     pub rl: RateLimit,
     pub dir: Directory,
+    pub key_type: KeyType,
+    pub signature_algorithm: JwsSignatureAlgorithm,
 }
 
 impl Endpoint {
@@ -25,8 +29,19 @@ impl Endpoint {
         url: &str,
         tos_agreed: bool,
         limits: &[(usize, String)],
+        key_type: &Option<String>,
+        signature_algorithm: &Option<String>,
     ) -> Result<Self, Error> {
         let rl = RateLimit::new(limits)?;
+        let key_type = match key_type {
+            Some(kt) => KeyType::from_str(&kt)?,
+            None => crate::DEFAULT_ACCOUNT_KEY_TYPE,
+        };
+        let signature_algorithm = match signature_algorithm {
+            Some(sa) => JwsSignatureAlgorithm::from_str(&sa)?,
+            None => key_type.get_default_signature_alg(),
+        };
+        let _ = key_type.check_alg_compatibility(&signature_algorithm)?;
         Ok(Self {
             name: name.to_string(),
             url: url.to_string(),
@@ -42,6 +57,8 @@ impl Endpoint {
                 revoke_cert: String::new(),
                 key_change: String::new(),
             },
+            key_type,
+            signature_algorithm,
         })
     }
 }